+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Fix an issue that introduced promises that would throw but the app could not handle.
 * Support for Koa.
 * Update dependencies.
+* Send one last stats report on process.exit() and SIGINT.
 
 ### 1.0.3
 * Support `HTTPS_PROXY` and a proxyUrl option.